Add Helpers and config files
to make the code more cleaner and maintainable add these files and do some adjustments to it.
