+++ src/js/config.js
@@ -1,0 +1,3 @@
+//Forkify API URL
+export const URL_API = `https://forkify-api.herokuapp.com/api/v2/recipes/`;
+export const TIMEOUT_SEC = 10;
